% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMX.R
\name{TablaAnova_MEM}
\alias{TablaAnova_MEM}
\title{Modelo con un efecto aleatorio y un efecto fijo}
\usage{
TablaAnova_MEM(respuesta, fijo, aleatorio, data)
}
\arguments{
\item{respuesta}{(variable respuesta) datos de la variable respuesta.}

\item{fijo}{(factor) datos del factor fijo.}

\item{aleatorio}{(factor) datos del factor aleatorio.}

\item{data}{(df) dataframes de los datos recolectados.}
}
\value{
una tabla ANOVA con la suma de los cuadrados entre y dentro de grupos, grados de libertad, la media de los cuadrados, la estadistica F y el valor p.
}
\description{
Realiza una tabla ANOVA para un modelo estadístico que contenga efectos fijos y aleatorios.
}
